Update to Simulation
A new parameter "unique_number" can now be given to the grid.set method.

If the parameter is None a new cell will be created with the unique_number incremented by one.

If the parameter is the unique_number of a cell this cell will be moved to the new position.

This allows the smoke and fire lifetime to work properly, as new smoke or fire cells are not being created every frame.