Significant performance increase
The get_by_unique_number function used to take 14s of CPU time (25 second render) and now only takes 0.5s (35 second render)

The function used to loop through all cells in the grid to find the cell with the unique_number, which is very inefficient. Now the Grid class has a dictionary that maps the cells to the unique_number.