Burning property and randomized grid simulation
Cells have a new property called 'burning,' which indicates whether a cell can set adjacent cells on fire.

Added new blocks 'fuse' and 'lit fuse' that burn quickly.

The addition of the fuse necessitated a randomized grid simulation. Previously, the grid was simulated from top to bottom and left to right, causing the fuse to burn instantly from bottom to top but taking the intended 1 frame from top to bottom.

With the randomized grid simulation, the fuse burns correctly in all directions, and as a bonus, the other blocks simulate more organically due to the random nature of the simulation.